pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voxelflex"
# Increment version for refactoring, consider using dynamic versioning later
version = "0.3.0"
authors = [
  # Replace with your actual name and email
  { name="Your Name", email="your.email@example.com" },
]
description = "Temperature-aware protein flexibility (RMSF) prediction from 3D voxel data using CNNs and a preprocessing workflow."
readme = "README.md"
requires-python = ">=3.9" # Based on type hints and f-strings
license = { file = "LICENSE" } # Add a LICENSE file (e.g., MIT)
classifiers = [
    "Development Status :: 3 - Alpha", # Update as project matures
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License", # Choose your license
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.21", # Check specific version needs
    "pandas>=1.3",
    "torch>=1.10.0", # Consider torch version carefully based on hardware/CUDA
    "scipy",
    "scikit-learn",
    "h5py",
    "pyyaml",
    "matplotlib", # Ensure compatible version if issues arise
    "seaborn",
    "psutil",
    "tqdm",
    # Add specific versions if needed for reproducibility
    # e.g., "torch==2.0.1", "pandas==2.0.3"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/voxelflex" # Replace with your repo URL
"Bug Tracker" = "https://github.com/yourusername/voxelflex/issues" # Replace

# Define the command-line script entry point
[project.scripts]
voxelflex = "voxelflex.cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

# Optional: Tool configurations (can be added later)
# [tool.black]
# line-length = 88
# target-version = ['py310']

# [tool.isort]
# profile = "black"

# [tool.mypy]
# python_version = "3.10"
# warn_return_any = true
# warn_unused_configs = true
# ignore_missing_imports = true # Initially, relax this if needed
